+++ src/irmin-cli/resolver.mli
@@ -86,7 +86,7 @@ module Store : sig
     t
 
   val mem : hash -> contents -> t
-  val irf : hash -> contents -> t
+  val fs : hash -> contents -> t
   val http : t -> t
   val git : contents -> t
   val pack : hash -> contents -> t